[UPDATE] add custom metadata and modify type of version from UInt to UInt64.

+++ Sources/DDDCore/Entity/AggregateRoot.swift
@@ -36,7 +36,7 @@ extension AggregateRoot {
         metadata.events
     }
 
-    public var version: UInt? {
+    public var version: UInt64? {
         metadata.version
     }
